packages/backend-workers/pyproject.toml
[project]
name = "musetric-backend-workers"
version = "0.1.0"
description = "Musetric Backend Workers"
authors = [{ name = "Vladlen Popelenkov", email = "popelenkow@gmail.com" }]
requires-python = "==3.12.6"
dependencies = [
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "rotary-embedding-torch==0.8.0",
    "einops==0.8.0",
    "ml-collections==0.1.1",
    "librosa==0.10.2",
    "soundfile==0.12.1",
    "numpy==1.26.4",
    "PyYAML==6.0.2",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "isort==6.0.1",
    "ruff==0.11.10"
]

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cuda" }
torchaudio = { index = "pytorch-cuda" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/musetric_backend_workers"]

[tool.black]
exclude = ".venv"

[tool.isort]
profile = "black"
skip = [".venv"]

[tool.ruff]
exclude = [".venv"]
lint.select = ["TID"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
